hack/requirements.txt
# Fast API framework and server
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Financial data APIs
sec-api>=1.0.17
requests>=2.28.0

# Data processing
pandas>=1.5.3
numpy>=1.24.0
beautifulsoup4>=4.12.0  # For HTML parsing in financial statements
lxml>=4.9.0  # Faster HTML parsing with BeautifulSoup
html5lib>=1.1  # Flexible HTML parsing

# Excel export
openpyxl>=3.1.0  # For creating Excel files
xlsxwriter>=3.0.9  # Alternative Excel writer with more formatting options
jinja2>=3.1.2  # For template-based Excel generation

# AI and NLP libraries
openai>=1.3.0  # Updated for vision model support
langchain>=0.0.200
anthropic>=0.5.0  # For Claude API access

# HTTP and utilities
python-dotenv>=1.0.0
aiohttp>=3.8.4  # For async HTTP requests
PyPDF2>=3.0.0  # For PDF processing
pytesseract>=0.3.10  # For OCR processing of PDFs
regex>=2023.0.0  # Enhanced regex capabilities for complex pattern matching

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Development
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
flake8>=6.0.0 